correction routes et modifyUser

+++ controllers/user.js
@@ -4,6 +4,7 @@ const User = require("../models/User");
 const validEmail = /^[a-zA-Z0-9+_.-]+@[a-zA-Z0-9.-]+$/;        //regex pour l'email. Normalement pris en charge par le front, mais mieux vaux double sécurité
 const validPassword = /^[^\s]{8,}$/;                           //regex pour le password. idem email.
 const fs = require("fs");
+const { crossOriginResourcePolicy } = require("helmet");
                                                                
 
 
@@ -15,6 +16,7 @@ exports.signup = (req, res, next) => {
     if (!req.body.password.match(validPassword)) {
         return res.status(400).json({ message: "le mot de passe doit contenir au moins 8 caracères" });
     }
+    console.log(req.body);
     User.findOne({ email: req.body.email })                                                                  //recherche si l'email correspond déjà à un compte
         .then(user => {
             if (user) {                                                                                      //sécurité : impossible de créer 2 comptes avec le même email
@@ -82,22 +84,23 @@ exports.getOneUser = (req, res, next) => {
     //.catch((error) => res.status(404).json({error}));
 };
 
-
+//TODO lien vers nouvelle image
 exports.modifyUser = (req, res, next) => {
     User.findOne({_id : req.params.id})
     .then(user => {
-        console.log(user);
-        // if(user._id != req.auth) {
-        //     return res.status(403).json({message : "ce compte n'est pas le vôtre, vous ne pouvez pas le modifier"});
-        // }
+        console.log(req.body);
+        if(user._id != req.auth.userId) {                   
+            return res.status(403).json({message : "ce compte n'est pas le vôtre, vous ne pouvez pas le modifier"});
+        }
         if (req.file) {
             fs.unlink(`images/${user.avatarUrl.split("/images/")[1]}`, () => {
-                User.updateOne({_id : req.params.id}, {...req.body, _id : req.params.id})
+                imageUrl = `${req.protocol}://${req.get("host")}/images/${req.file.filename}`;
+                User.updateOne({_id : req.params.id}, {...req.body})                                     
                 .then(() => res.status(200).json({message : "compte modifié"}))
              //   .catch((error) => res.status(500).json({error}));
             })
         } else {
-            User.updateOne({_id : req.params.id}, {...req.body, _id : req.params.id})
+            User.updateOne({_id : req.params.id}, {...req.body})
             .then(() => res.status(200).json({message : "compte modifié"}))
           //  .catch((error) => res.status(500).json({error}));
         }
